move SMS Gateway Fee Criteria interface to smsbillables
