+++ corehq/apps/smsbillables/interface.py
@@ -16,7 +16,10 @@ from corehq.apps.smsbillables.filters import (
     DomainFilter,
     ShowBillablesFilter,
 )
-from corehq.apps.smsbillables.models import SmsBillable
+from corehq.apps.smsbillables.models import (
+    SmsBillable,
+    SmsGatewayFeeCriteria,
+)
 
 
 class SMSBillablesInterface(GenericTabularReport):
@@ -98,3 +101,43 @@ class SMSBillablesInterface(GenericTabularReport):
                 domain=domain,
             )
         return selected_billables
+
+
+class SMSGatewayFeeCriteriaInterface(GenericTabularReport):
+    base_template = "accounting/report_filter_actions.html"
+    section_name = "Accounting"
+    dispatcher = AccountingAdminInterfaceDispatcher
+    name = "SMS Gateway Fee Criteria"
+    description = "List of all SMS Gateway Fee Criteria"
+    slug = "sms_gateway_fee_criteria"
+    fields = [
+    ]
+
+    @property
+    def headers(self):
+        return DataTablesHeader(
+            DataTablesColumn("Gateway Type"),
+            DataTablesColumn("Specific Gateway"),
+            DataTablesColumn("Direction"),
+            DataTablesColumn("Country Code"),
+        )
+
+    @property
+    def rows(self):
+        return [
+            [
+                criteria.backend_api_id,
+                (criteria.backend_instance
+                 if criteria.backend_instance is not None else "Any"),
+                criteria.direction,
+                (criteria.country_code
+                 if criteria.country_code is not None else "Any"),
+            ]
+            for criteria in self.sms_gateway_fee_criteria
+        ]
+
+    @property
+    def sms_gateway_fee_criteria(self):
+        selected_criteria = SmsGatewayFeeCriteria.objects
+        selected_criteria = selected_criteria.filter()  # TODO - apply filters
+        return selected_criteria
